Split frontend JSON symbol tests
